Refactor: split raw profile internals into focused modules (#122)
