+++ toolkit/profile/__init__.py
@@ -5,6 +5,7 @@ from toolkit.profile.raw import (
     build_profile_hints,
     build_suggested_read_cfg,
     profile_raw,
+    write_raw_profile,
     write_suggested_read_yml,
 )
 
@@ -13,5 +14,6 @@ __all__ = [
     "build_profile_hints",
     "build_suggested_read_cfg",
     "profile_raw",
+    "write_raw_profile",
     "write_suggested_read_yml",
 ]
